requirements.txt
fastapi
uvicorn
streamlit
pyttsx3
openai-whisper
openai
crewai
yfinance
bs4
requests
langchain
faiss-cpu
sentence-transformers
